Speech-LLM/requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.6.0
rich>=13.7.0

# Speech
openai-whisper>=20231117
gTTS>=2.5.0

# LLM backends (optional)
openai>=1.3.0
transformers>=4.41.0
accelerate>=0.30.0
sentencepiece>=0.2.0

# Multi-agent
crewai>=0.64.0
